Moved StaffProfile.js to component folder.
Deleted Screens folder.
