--- Screens/StaffView.js
@@ -1,21 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-
-export default function StaffProfile() {
-    return (
-      <View style={styles.container}>
-        <Text>Jacob created this Staff View.</Text>
-        <StatusBar style="auto" />
-      </View>
-    );
-  }
-  
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-  });
